@@ -27,39 +27,39 @@
 		"@types/oauth": "0.9.1",
 		"@types/parse5": "6.0.3",
 		"@types/punycode": "2.1.0",
-		"@types/qrcode": "1.4.1",
+		"@types/qrcode": "1.4.2",
 		"@types/random-seed": "0.3.3",
 		"@types/request-stats": "3.0.0",
 		"@types/seedrandom": "2.4.28",
 		"@types/throttle-debounce": "2.1.0",
 		"@types/tinycolor2": "1.4.3",
-		"@types/tmp": "0.2.2",
+		"@types/tmp": "0.2.3",
 		"@types/uuid": "8.3.3",
 		"@types/web-push": "3.3.2",
 		"@types/webpack": "5.28.0",
 		"@types/webpack-stream": "3.2.12",
 		"@types/websocket": "1.0.4",
 		"@types/ws": "8.2.2",
-		"@typescript-eslint/parser": "5.6.0",
-		"@vue/compiler-sfc": "3.2.24",
+		"@typescript-eslint/parser": "5.8.1",
+		"@vue/compiler-sfc": "3.2.26",
 		"abort-controller": "3.0.0",
 		"autobind-decorator": "2.4.0",
 		"autosize": "4.0.4",
 		"autwh": "0.1.0",
 		"blurhash": "1.1.4",
-		"broadcast-channel": "4.7.0",
-		"chart.js": "3.6.2",
+		"broadcast-channel": "4.9.0",
+		"chart.js": "3.7.0",
 		"chartjs-adapter-date-fns": "2.0.0",
 		"chartjs-plugin-zoom": "1.2.0",
 		"compare-versions": "3.6.0",
-		"content-disposition": "0.5.3",
+		"content-disposition": "0.5.4",
 		"crc-32": "1.2.0",
 		"css-loader": "6.5.1",
-		"cssnano": "5.0.12",
-		"date-fns": "2.27.0",
+		"cssnano": "5.0.14",
+		"date-fns": "2.28.0",
 		"dateformat": "4.5.1",
 		"escape-regexp": "0.0.1",
-		"eslint": "8.4.1",
+		"eslint": "8.5.0",
 		"eslint-plugin-vue": "8.2.0",
 		"eventemitter3": "4.0.7",
 		"feed": "4.2.2",
@@ -80,7 +80,7 @@
 		"parse5": "6.0.1",
 		"photoswipe": "git://github.com/dimsemenov/photoswipe#v5-beta",
 		"portscanner": "2.2.0",
-		"postcss": "8.4.4",
+		"postcss": "8.4.5",
 		"postcss-loader": "6.2.1",
 		"prismjs": "1.25.0",
 		"private-ip": "2.3.3",
@@ -94,7 +94,7 @@
 		"request-stats": "3.0.0",
 		"rndstr": "1.0.0",
 		"s-age": "1.1.2",
-		"sass": "1.44.0",
+		"sass": "1.45.1",
 		"sass-loader": "12.4.0",
 		"seedrandom": "3.0.5",
 		"strict-event-emitter-types": "2.0.0",
@@ -111,11 +111,11 @@
 		"tsc-alias": "1.4.2",
 		"tsconfig-paths": "3.12.0",
 		"twemoji-parser": "13.1.0",
-		"typescript": "4.5.2",
+		"typescript": "4.5.4",
 		"uuid": "8.3.2",
 		"v-debounce": "0.1.2",
 		"vanilla-tilt": "1.7.2",
-		"vue": "3.2.24",
+		"vue": "3.2.26",
 		"vue-loader": "16.8.3",
 		"vue-prism-editor": "2.0.0-alpha.2",
 		"vue-router": "4.0.5",
@@ -126,14 +126,14 @@
 		"webpack": "5.65.0",
 		"webpack-cli": "4.9.1",
 		"websocket": "1.0.34",
-		"ws": "8.3.0"
+		"ws": "8.4.0"
 	},
 	"devDependencies": {
 		"@redocly/openapi-core": "1.0.0-beta.54",
 		"@types/fluent-ffmpeg": "2.1.17",
-		"@typescript-eslint/eslint-plugin": "5.4.0",
+		"@typescript-eslint/eslint-plugin": "5.8.1",
 		"cross-env": "7.0.3",
-		"cypress": "8.5.0",
+		"cypress": "9.2.0",
 		"eslint-plugin-import": "2.25.3",
 		"start-server-and-test": "1.14.0"
 	}
