@@ -3,7 +3,7 @@ import { App } from 'vue';
 import userPreview from './user-preview';
 import size from './size';
 import getSize from './get-size';
-import particle from './particle';
+import ripple from './ripple';
 import tooltip from './tooltip';
 import hotkey from './hotkey';
 import appear from './appear';
@@ -11,13 +11,14 @@ import anim from './anim';
 import stickyContainer from './sticky-container';
 import clickAnime from './click-anime';
 import panel from './panel';
+import adaptiveBorder from './adaptive-border';
 
 export default function(app: App) {
 	app.directive('userPreview', userPreview);
 	app.directive('user-preview', userPreview);
 	app.directive('size', size);
 	app.directive('get-size', getSize);
-	app.directive('particle', particle);
+	app.directive('ripple', ripple);
 	app.directive('tooltip', tooltip);
 	app.directive('hotkey', hotkey);
 	app.directive('appear', appear);
@@ -25,4 +26,5 @@ export default function(app: App) {
 	app.directive('click-anime', clickAnime);
 	app.directive('sticky-container', stickyContainer);
 	app.directive('panel', panel);
+	app.directive('adaptive-border', adaptiveBorder);
 }
